+++ lib/chain.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const doInject = require('./doInject')
+const doInject = require('./do-inject')
 
 const errorMessage = 'The dispatch function has already been invoked'
