+++ lib/do-inject.js
